version final/proyecto aseguramiento/requirements.txt
Flask
pymongo
werkzeug
certifi
python-dotenv
pytest
mongomock
coverage